Migrate codebase to Typescript
