--- extras/index.js
@@ -1,1 +1,0 @@
-export { MiniStats } from './mini-stats/mini-stats.js';
