+++ extras/index.ts
@@ -1,0 +1,1 @@
+export { MiniStats } from './mini-stats/mini-stats';
