+++ extras/mini-stats/cpu-timer.ts
@@ -1,21 +1,28 @@
-function CpuTimer(app) {
-    this._frameIndex = 0;
-    this._frameTimings = [];
-    this._timings = [];
-    this._prevTimings = [];
-    this.unitsName = "ms";
-    this.decimalPlaces = 1;
+class CpuTimer {
+    _frameIndex: number;
+    _frameTimings: [string, number][];
+    _timings: [string, number][];
+    _prevTimings: [string, number][];
+    unitsName: string;
+    decimalPlaces: number;
+    enabled: boolean;
 
-    this.enabled = true;
+    constructor(app: pc.Application) {
+        this._frameIndex = 0;
+        this._frameTimings = [];
+        this._timings = [];
+        this._prevTimings = [];
+        this.unitsName = "ms";
+        this.decimalPlaces = 1;
+        this.enabled = true;
 
-    app.on('frameupdate', this.begin.bind(this, 'update'));
-    app.on('framerender', this.mark.bind(this, 'render'));
-    app.on('frameend', this.mark.bind(this, 'other'));
-}
+        app.on('frameupdate', this.begin.bind(this, 'update'));
+        app.on('framerender', this.mark.bind(this, 'render'));
+        app.on('frameend', this.mark.bind(this, 'other'));
+    }
 
-Object.assign(CpuTimer.prototype, {
     // mark the beginning of the frame
-    begin: function (name) {
+    begin(name: string): void {
         if (!this.enabled) {
             return;
         }
@@ -31,10 +38,10 @@ Object.assign(CpuTimer.prototype, {
         this._frameIndex = 0;
 
         this.mark(name);
-    },
+    }
 
     // mark
-    mark: function (name) {
+    mark(name: string): void {
         if (!this.enabled) {
             return;
         }
@@ -60,16 +67,14 @@ Object.assign(CpuTimer.prototype, {
         }
         this._frameIndex++;
     }
-});
 
-Object.defineProperty(CpuTimer.prototype, 'timings', {
-    get: function () {
+    get timings(): number[] {
         // remove the last time point from the list (which is the time spent outside
         // of playcanvas)
         return this._timings.slice(0, -1).map(function (v) {
             return v[1];
         });
     }
-});
+}
 
 export { CpuTimer };
