+++ extras/mini-stats/gpu-timer.ts
@@ -1,26 +1,42 @@
-function GpuTimer(app) {
-    this._gl = app.graphicsDevice.gl;
-    this._ext = app.graphicsDevice.extDisjointTimerQuery;
-
-    this._freeQueries = [];                     // pool of free queries
-    this._frameQueries = [];                    // current frame's queries
-    this._frames = [];                          // list of previous frame queries
-
-    this._timings = [];
-    this._prevTimings = [];
-
-    this.enabled = true;
-    this.unitsName = "ms";
-    this.decimalPlaces = 1;
-
-    app.on('frameupdate', this.begin.bind(this, 'update'));
-    app.on('framerender', this.mark.bind(this, 'render'));
-    app.on('frameend', this.end.bind(this));
+interface EXTDisjointTimerQueryWebGL {
+    GPU_DISJOINT_EXT: number,
+    TIME_ELAPSED_EXT: number
 }
 
-Object.assign(GpuTimer.prototype, {
+class GpuTimer {
+    _gl: WebGL2RenderingContext;
+    _ext: EXTDisjointTimerQueryWebGL;
+    _freeQueries: [string, WebGLQuery][];
+    _frameQueries: [string, WebGLQuery][];
+    _frames: [string, WebGLQuery][][];
+    _timings: [string, number][];
+    _prevTimings: [string, number][];
+    enabled: boolean;
+    unitsName: string;
+    decimalPlaces: number;
+
+    constructor(app: pc.Application) {
+        this._gl = app.graphicsDevice.gl;
+        this._ext = app.graphicsDevice.extDisjointTimerQuery;
+
+        this._freeQueries = [];                     // pool of free queries
+        this._frameQueries = [];                    // current frame's queries
+        this._frames = [];                          // list of previous frame queries
+
+        this._timings = [];
+        this._prevTimings = [];
+
+        this.enabled = true;
+        this.unitsName = "ms";
+        this.decimalPlaces = 1;
+
+        app.on('frameupdate', this.begin.bind(this, 'update'));
+        app.on('framerender', this.mark.bind(this, 'render'));
+        app.on('frameend', this.end.bind(this));
+    }
+
     // mark the beginning of the frame
-    begin: function (name) {
+    begin(name: string): void {
         if (!this.enabled) {
             return;
         }
@@ -47,10 +63,10 @@ Object.assign(GpuTimer.prototype, {
         }
 
         this.mark(name);
-    },
+    }
 
     // mark
-    mark: function (name) {
+    mark(name: string): void {
         if (!this.enabled) {
             return;
         }
@@ -65,10 +81,10 @@ Object.assign(GpuTimer.prototype, {
         query[0] = name;
         this._gl.beginQuery(this._ext.TIME_ELAPSED_EXT, query[1]);
         this._frameQueries.push(query);
-    },
+    }
 
     // end of frame
-    end: function () {
+    end(): void {
         if (!this.enabled) {
             return;
         }
@@ -76,11 +92,11 @@ Object.assign(GpuTimer.prototype, {
         this._gl.endQuery(this._ext.TIME_ELAPSED_EXT);
         this._frames.push(this._frameQueries);
         this._frameQueries = [];
-    },
+    }
 
     // check if the gpu has been interrupted thereby invalidating all
     // in-flight queries
-    _checkDisjoint: function () {
+    _checkDisjoint(): void {
         var disjoint = this._gl.getParameter(this._ext.GPU_DISJOINT_EXT);
         if (disjoint) {
             // return all queries to the free list
@@ -88,16 +104,16 @@ Object.assign(GpuTimer.prototype, {
             this._frameQueries = [];
             this._frames = [];
         }
-    },
+    }
 
     // either returns a previously free'd query or if there aren't any allocates a new one
-    _allocateQuery: function () {
+    _allocateQuery(): [string, WebGLQuery] {
         return (this._freeQueries.length > 0) ?
             this._freeQueries.splice(-1, 1)[0] : ["", this._gl.createQuery()];
-    },
+    }
 
     // attempt to resolve one frame's worth of timings
-    _resolveFrameTimings: function (frame, timings) {
+    _resolveFrameTimings(frame, timings): boolean {
         // wait for the last query in the frame to be available
         if (!this._gl.getQueryParameter(frame[frame.length - 1][1], this._gl.QUERY_RESULT_AVAILABLE)) {
             return false;
@@ -109,14 +125,12 @@ Object.assign(GpuTimer.prototype, {
 
         return true;
     }
-});
 
-Object.defineProperty(GpuTimer.prototype, 'timings', {
-    get: function () {
+    get timings() {
         return this._timings.map(function (v) {
             return v[1];
         });
     }
-});
+}
 
 export { GpuTimer };
