+++ package.json
@@ -1,74 +1,77 @@
 {
-  "name": "playcanvas", 
-  "version": "1.37.0-dev", 
-  "author": "PlayCanvas <support@playcanvas.com>", 
-  "homepage": "https://playcanvas.com", 
-  "description": "PlayCanvas WebGL game engine", 
+  "name": "playcanvas",
+  "version": "1.37.0-dev",
+  "author": "PlayCanvas <support@playcanvas.com>",
+  "homepage": "https://playcanvas.com",
+  "description": "PlayCanvas WebGL game engine",
   "keywords": [
-    "3D", 
-    "2D", 
-    "VR", 
-    "WebGL", 
-    "WebGL2", 
-    "game", 
-    "engine", 
-    "HTML5", 
-    "browser", 
+    "3D",
+    "2D",
+    "VR",
+    "WebGL",
+    "WebGL2",
+    "game",
+    "engine",
+    "HTML5",
+    "browser",
     "typings"
-  ], 
-  "license": "MIT", 
-  "main": "build/playcanvas.js", 
-  "types": "index.d.ts", 
+  ],
+  "license": "MIT",
+  "main": "build/playcanvas.js",
+  "types": "index.d.ts",
   "bugs": {
     "url": "https://github.com/playcanvas/engine/issues"
-  }, 
+  },
   "repository": {
-    "type": "git", 
+    "type": "git",
     "url": "https://github.com/playcanvas/engine.git"
-  }, 
+  },
   "eslintConfig": {
     "extends": "@playcanvas/eslint-config"
-  }, 
+  },
   "eslintIgnore": [
-    "examples/lib/*", 
+    "examples/lib/*",
     "src/polyfill/*"
-  ], 
+  ],
   "devDependencies": {
-    "chai": "^4.2.0", 
-    "eslint": "^7.13.0", 
-    "google-closure-compiler": "20201102.0.1", 
-    "http-server": "^0.12.3", 
-    "jsdoc": "^3.6.6", 
-    "karma": "^5.2.3", 
-    "karma-chrome-launcher": "^3.1.0", 
-    "karma-mocha": "^2.0.1", 
-    "karma-spec-reporter": "0.0.32", 
-    "mocha": "^8.2.1", 
-    "@playcanvas/eslint-config": "^1.0.6", 
-    "@playcanvas/jsdoc-template": "^1.0.13", 
-    "preprocessor": "^1.4.0", 
-    "rollup": "^2.33.1", 
-    "@rollup/plugin-replace": "^2.3.4", 
-    "@rollup/pluginutils": "^4.1.0", 
-    "rollup-plugin-cleanup": "^3.2.1", 
-    "sinon": "^9.2.1", 
-    "tsd-jsdoc": "^2.5.0", 
-    "typescript": "^4.0.5", 
+    "chai": "^4.2.0",
+    "eslint": "^7.13.0",
+    "google-closure-compiler": "20201102.0.1",
+    "http-server": "^0.12.3",
+    "jsdoc": "^3.6.6",
+    "karma": "^5.2.3",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-mocha": "^2.0.1",
+    "karma-spec-reporter": "0.0.32",
+    "mocha": "^8.2.1",
+    "@playcanvas/eslint-config": "^1.0.6",
+    "@playcanvas/jsdoc-template": "^1.0.13",
+    "preprocessor": "^1.4.0",
+    "rollup": "^2.33.1",
+    "@rollup/plugin-replace": "^2.3.4",
+    "@rollup/plugin-typescript": "^6.1.0",
+    "@rollup/pluginutils": "^4.1.0",
+    "rollup-plugin-cleanup": "^3.2.1",
+    "sinon": "^9.2.1",
+    "tslib": "^2.0.3",
+    "tsd-jsdoc": "^2.5.0",
+    "typescript": "^4.0.5",
+    "@types/webgl2": "0.0.5",
     "uglify-js": "^3.11.5"
-  }, 
+  },
   "scripts": {
-    "build": "rollup -c", 
-    "build:all": "npm run build && npm run closure", 
-    "closure": "google-closure-compiler --compilation_level=SIMPLE --warning_level=VERBOSE --jscomp_off=nonStandardJsDocs --jscomp_off=checkTypes --externs externs.js --language_in=ECMASCRIPT5_STRICT --js build/playcanvas.js --js_output_file build/playcanvas.min.js", 
-    "docs": "jsdoc -c conf-api.json", 
-    "lint": "eslint --ext .js examples extras src rollup.config.js", 
-    "serve": "http-server build -a localhost -p 51000", 
-    "test": "karma start tests/karma.conf.js -- --single-run --release", 
-    "test:watch": "karma start tests/karma.conf.js", 
-    "test:debug": "karma start tests/karma.conf.js -- --single-run=false", 
-    "test:tsd": "npm run tsd && tsc --pretty false build/playcanvas.d.ts", 
-    "tsd": "jsdoc -c conf-tsd.json", 
-    "uglify": "uglifyjs build/playcanvas.js --compress --mangle --output build/playcanvas.min.js", 
+    "build": "rollup -c",
+    "build:all": "npm run build && npm run closure",
+    "closure": "google-closure-compiler --compilation_level=SIMPLE --warning_level=VERBOSE --jscomp_off=nonStandardJsDocs --jscomp_off=checkTypes --externs externs.js --language_in=ECMASCRIPT5_STRICT --js build/playcanvas.js --js_output_file build/playcanvas.min.js",
+    "docs": "jsdoc -c conf-api.json",
+    "lint": "eslint --ext .js examples extras src rollup.config.js",
+    "serve": "http-server build -a localhost -p 51000",
+    "test": "karma start tests/karma.conf.js -- --single-run --release",
+    "test:watch": "karma start tests/karma.conf.js",
+    "test:debug": "karma start tests/karma.conf.js -- --single-run=false",
+    "test:tsd": "npm run tsd && tsc --pretty false build/playcanvas.d.ts",
+    "tsd": "jsdoc -c conf-tsd.json",
+    "uglify": "uglifyjs build/playcanvas.js --compress --mangle --output build/playcanvas.min.js",
     "unzipbundle": "gzip -x tests/assets/bundle.gz"
   }
 }
