+++ src/math/curve-evaluator.ts
@@ -1,23 +1,32 @@
-import { CURVE_CARDINAL, CURVE_CATMULL, CURVE_LINEAR, CURVE_SMOOTHSTEP, CURVE_SPLINE, CURVE_STEP } from './constants.js';
-import { math } from './math.js';
+import { CURVE_CARDINAL, CURVE_CATMULL, CURVE_LINEAR, CURVE_SMOOTHSTEP, CURVE_SPLINE, CURVE_STEP } from './constants';
+import { math } from './math';
+import { Curve } from './curve';
 
-function CurveEvaluator(curve, time) {
-    this._curve = curve;
-    this._left = -Infinity;
-    this._right = Infinity;
-    this._recip = 0;
-    this._p0 = 0;
-    this._p1 = 0;
-    this._m0 = 0;
-    this._m1 = 0;
-    this._reset(time || 0);
-}
+class CurveEvaluator {
+    _curve: Curve;
+    _left: number;
+    _right: number;
+    _recip: number;
+    _p0: number;
+    _p1: number;
+    _m0: number;
+    _m1: number;
 
-Object.assign(CurveEvaluator.prototype, {
+    constructor(curve: Curve, time?: number) {
+        this._curve = curve;
+        this._left = -Infinity;
+        this._right = Infinity;
+        this._recip = 0;
+        this._p0 = 0;
+        this._p1 = 0;
+        this._m0 = 0;
+        this._m1 = 0;
+        this._reset(time || 0);
+    }
 
     // Evaluate the curve at the given time. Specify forceReset if the
     // underlying curve keys have changed since the last evaluation.
-    evaluate: function (time, forceReset) {
+    evaluate(time: number, forceReset?: boolean): number {
         if (forceReset || time < this._left || time >= this._right) {
             this._reset(time);
         }
@@ -44,10 +53,10 @@ Object.assign(CurveEvaluator.prototype, {
             }
         }
         return result;
-    },
+    }
 
     // Calculate weights for the curve interval at the given time
-    _reset: function (time) {
+    _reset(time: number): void {
         var keys = this._curve.keys;
         var len = keys.length;
 
@@ -93,17 +102,17 @@ Object.assign(CurveEvaluator.prototype, {
                 }
             }
         }
-    },
+    }
 
     // returns true if the curve is a hermite and false otherwise
-    _isHermite: function () {
+    _isHermite(): boolean {
         return this._curve.type === CURVE_CATMULL ||
                this._curve.type === CURVE_CARDINAL ||
                this._curve.type === CURVE_SPLINE;
-    },
+    }
 
     // calculate tangents for the hermite curve
-    _calcTangents: function (keys, index) {
+    _calcTangents(keys: Array<number[]>, index: number): void {
         var a;
         var b = keys[index];
         var c = keys[index + 1];
@@ -143,9 +152,9 @@ Object.assign(CurveEvaluator.prototype, {
             this._m0 = tension * (c[1] - a_);
             this._m1 = tension * (d_ - b[1]);
         }
-    },
+    }
 
-    _evaluateHermite: function (p0, p1, m0, m1, t) {
+    _evaluateHermite(p0: number, p1: number, m0: number, m1: number, t: number): number {
         var t2 = t * t;
         var twot = t + t;
         var omt = 1 - t;
@@ -155,6 +164,6 @@ Object.assign(CurveEvaluator.prototype, {
                p1 * (t2 * (3 - twot)) +
                m1 * (t2 * (t - 1));
     }
-});
+}
 
 export { CurveEvaluator };
