+++ src/math/curve-set.ts
@@ -1,6 +1,6 @@
-import { CURVE_SMOOTHSTEP } from './constants.js';
-import { Curve } from './curve.js';
-import { CurveEvaluator } from './curve-evaluator.js';
+import { CURVE_SMOOTHSTEP } from './constants';
+import { Curve } from './curve';
+import { CurveEvaluator } from './curve-evaluator';
 
 /**
  * @class
@@ -25,35 +25,38 @@ import { CurveEvaluator } from './curve-evaluator.js';
  *     ]
  * ]);
  */
-function CurveSet() {
-    var i;
+class CurveSet {
+    curves: Curve[];
+    _type: number;
 
-    this.curves = [];
-    this._type = CURVE_SMOOTHSTEP;
+    constructor(curveKeys?: number[][]) {
+        var i;
 
-    if (arguments.length > 1) {
-        for (i = 0; i < arguments.length; i++) {
-            this.curves.push(new Curve(arguments[i]));
-        }
-    } else {
-        if (arguments.length === 0) {
-            this.curves.push(new Curve());
+        this.curves = [];
+        this._type = CURVE_SMOOTHSTEP;
+
+        if (arguments.length > 1) {
+            for (i = 0; i < arguments.length; i++) {
+                this.curves.push(new Curve(arguments[i]));
+            }
         } else {
-            var arg = arguments[0];
-            if (typeof arg === 'number') {
-                for (i = 0; i < arg; i++) {
-                    this.curves.push(new Curve());
-                }
+            if (arguments.length === 0) {
+                this.curves.push(new Curve());
             } else {
-                for (i = 0; i < arg.length; i++) {
-                    this.curves.push(new Curve(arg[i]));
+                var arg = arguments[0];
+                if (typeof arg === 'number') {
+                    for (i = 0; i < arg; i++) {
+                        this.curves.push(new Curve());
+                    }
+                } else {
+                    for (i = 0; i < arg.length; i++) {
+                        this.curves.push(new Curve(arg[i]));
+                    }
                 }
             }
         }
     }
-}
 
-Object.assign(CurveSet.prototype, {
     /**
      * @function
      * @name pc.CurveSet#get
@@ -61,9 +64,9 @@ Object.assign(CurveSet.prototype, {
      * @param {number} index - The index of the curve to return.
      * @returns {pc.Curve} The curve at the specified index.
      */
-    get: function (index) {
+    get(index: number): Curve {
         return this.curves[index];
-    },
+    }
 
     /**
      * @function
@@ -76,7 +79,7 @@ Object.assign(CurveSet.prototype, {
      * to return the result.
      * @returns {number[]} The interpolated curve values at the specified time.
      */
-    value: function (time, result) {
+    value(time: number, result: number[]): number[] {
         var length = this.curves.length;
         result = result || [];
         result.length = length;
@@ -86,7 +89,7 @@ Object.assign(CurveSet.prototype, {
         }
 
         return result;
-    },
+    }
 
     /**
      * @function
@@ -94,7 +97,7 @@ Object.assign(CurveSet.prototype, {
      * @description Returns a clone of the specified curve set object.
      * @returns {pc.CurveSet} A clone of the specified curve set.
      */
-    clone: function () {
+    clone(): CurveSet {
         var result = new CurveSet();
 
         result.curves = [];
@@ -105,9 +108,9 @@ Object.assign(CurveSet.prototype, {
         result._type = this._type;
 
         return result;
-    },
+    }
 
-    quantize: function (precision) {
+    quantize(precision: number): Float32Array {
         precision = Math.max(precision, 2);
 
         var numCurves = this.curves.length;
@@ -122,7 +125,7 @@ Object.assign(CurveSet.prototype, {
         }
 
         return values;
-    },
+    }
 
     /**
      * @private
@@ -133,52 +136,48 @@ Object.assign(CurveSet.prototype, {
      * @param {number} precision - The number of samples to return.
      * @param {number} min - The minimum output value.
      * @param {number} max - The maximum output value.
-     * @returns {number[]} The set of quantized values.
+     * @returns {Float32Array} The set of quantized values.
      */
-    quantizeClamped: function (precision, min, max) {
+    quantizeClamped(precision: number, min: number, max: number): Float32Array {
         var result = this.quantize(precision);
         for (var i = 0; i < result.length; ++i) {
             result[i] = Math.min(max, Math.max(min, result[i]));
         }
         return result;
     }
-});
 
-/**
- * @readonly
- * @name pc.CurveSet#length
- * @type {number}
- * @description The number of curves in the curve set.
- */
-Object.defineProperty(CurveSet.prototype, 'length', {
-    get: function () {
+    /**
+     * @readonly
+     * @name pc.CurveSet#length
+     * @type {number}
+     * @description The number of curves in the curve set.
+     */
+    get length(): number {
         return this.curves.length;
     }
-});
 
-/**
- * @name pc.CurveSet#type
- * @type {number}
- * @description The interpolation scheme applied to all curves in the curve set. Can be:
- *
- * * {@link pc.CURVE_LINEAR}
- * * {@link pc.CURVE_SMOOTHSTEP}
- * * {@link pc.CURVE_SPLINE}
- * * {@link pc.CURVE_STEP}
- *
- * Defaults to {@link pc.CURVE_SMOOTHSTEP}.
- */
-Object.defineProperty(CurveSet.prototype, 'type', {
-    get: function () {
+    /**
+     * @name pc.CurveSet#type
+     * @type {number}
+     * @description The interpolation scheme applied to all curves in the curve set. Can be:
+     *
+     * * {@link pc.CURVE_LINEAR}
+     * * {@link pc.CURVE_SMOOTHSTEP}
+     * * {@link pc.CURVE_SPLINE}
+     * * {@link pc.CURVE_STEP}
+     *
+     * Defaults to {@link pc.CURVE_SMOOTHSTEP}.
+     */
+    get type(): number {
         return this._type;
-    },
+    }
 
-    set: function (value) {
+    set type(value: number) {
         this._type = value;
         for (var i = 0; i < this.curves.length; i++) {
             this.curves[i].type = value;
         }
     }
-});
+}
 
 export { CurveSet };
