+++ src/math/index.ts
@@ -1,0 +1,10 @@
+export * from './constants';
+export { math } from './math';
+export { Curve } from './curve';
+export { CurveSet } from './curve-set';
+export { Mat3 } from './mat3';
+export { Mat4 } from './mat4';
+export { Quat } from './quat';
+export { Vec2 } from './vec2';
+export { Vec3 } from './vec3';
+export { Vec4 } from './vec4';
