+++ src/math/vec4.ts
@@ -10,21 +10,78 @@
  * @example
  * var v = new pc.Vec4(1, 2, 3, 4);
  */
-function Vec4(x, y, z, w) {
-    if (x && x.length === 4) {
-        this.x = x[0];
-        this.y = x[1];
-        this.z = x[2];
-        this.w = x[3];
-    } else {
-        this.x = x || 0;
-        this.y = y || 0;
-        this.z = z || 0;
-        this.w = w || 0;
-    }
-}
+class Vec4 {
+    /**
+     * @name pc.Vec4#x
+     * @type {number}
+     * @description The first component of the vector.
+     * @example
+     * var vec = new pc.Vec4(10, 20, 30, 40);
+     *
+     * // Get x
+     * var x = vec.x;
+     *
+     * // Set x
+     * vec.x = 0;
+     */
+    x: number;
+    /**
+     * @name pc.Vec4#y
+     * @type {number}
+     * @description The second component of the vector.
+     * @example
+     * var vec = new pc.Vec4(10, 20, 30, 40);
+     *
+     * // Get y
+     * var y = vec.y;
+     *
+     * // Set y
+     * vec.y = 0;
+     */
+    y: number;
+    /**
+     * @name pc.Vec4#z
+     * @type {number}
+     * @description The third component of the vector.
+     * @example
+     * var vec = new pc.Vec4(10, 20, 30, 40);
+     *
+     * // Get z
+     * var z = vec.z;
+     *
+     * // Set z
+     * vec.z = 0;
+     */
+    z: number;
+    /**
+     * @name pc.Vec4#w
+     * @type {number}
+     * @description The fourth component of the vector.
+     * @example
+     * var vec = new pc.Vec4(10, 20, 30, 40);
+     *
+     * // Get z
+     * var z = vec.z;
+     *
+     * // Set z
+     * vec.z = 0;
+     */
+    w: number;
 
-Object.assign(Vec4.prototype, {
+    constructor(x?: number | number[], y?: number, z?: number, w?: number) {
+        if (Array.isArray(x)) {
+            this.x = x[0];
+            this.y = x[1];
+            this.z = x[2];
+            this.w = x[3];
+        } else {
+            this.x = x || 0;
+            this.y = y || 0;
+            this.z = z || 0;
+            this.w = w || 0;
+        }
+    }
+
     /**
      * @function
      * @name pc.Vec4#add
@@ -40,14 +97,14 @@ Object.assign(Vec4.prototype, {
      * // Should output [30, 30, 30]
      * console.log("The result of the addition is: " + a.toString());
      */
-    add: function (rhs) {
+    add(rhs: Vec4): Vec4 {
         this.x += rhs.x;
         this.y += rhs.y;
         this.z += rhs.z;
         this.w += rhs.w;
 
         return this;
-    },
+    }
 
     /**
      * @function
@@ -66,14 +123,14 @@ Object.assign(Vec4.prototype, {
      *
      * console.log("The result of the addition is: " + r.toString());
      */
-    add2: function (lhs, rhs) {
+    add2(lhs: Vec4, rhs: Vec4): Vec4 {
         this.x = lhs.x + rhs.x;
         this.y = lhs.y + rhs.y;
         this.z = lhs.z + rhs.z;
         this.w = lhs.w + rhs.w;
 
         return this;
-    },
+    }
 
     /**
      * @function
@@ -85,9 +142,9 @@ Object.assign(Vec4.prototype, {
      * var vclone = v.clone();
      * console.log("The result of the cloning is: " + vclone.toString());
      */
-    clone: function () {
+    clone(): Vec4 {
         return new Vec4().copy(this);
-    },
+    }
 
     /**
      * @function
@@ -103,14 +160,14 @@ Object.assign(Vec4.prototype, {
      *
      * console.log("The two vectors are " + (dst.equals(src) ? "equal" : "different"));
      */
-    copy: function (rhs) {
+    copy(rhs: Vec4): Vec4 {
         this.x = rhs.x;
         this.y = rhs.y;
         this.z = rhs.z;
         this.w = rhs.w;
 
         return this;
-    },
+    }
 
     /**
      * @function
@@ -124,9 +181,9 @@ Object.assign(Vec4.prototype, {
      * var v1dotv2 = v1.dot(v2);
      * console.log("The result of the dot product is: " + v1dotv2);
      */
-    dot: function (rhs) {
+    dot(rhs: Vec4): number {
         return this.x * rhs.x + this.y * rhs.y + this.z * rhs.z + this.w * rhs.w;
-    },
+    }
 
     /**
      * @function
@@ -139,9 +196,9 @@ Object.assign(Vec4.prototype, {
      * var b = new pc.Vec4(5, 6, 7, 8);
      * console.log("The two vectors are " + (a.equals(b) ? "equal" : "different"));
      */
-    equals: function (rhs) {
+    equals(rhs: Vec4): boolean {
         return this.x === rhs.x && this.y === rhs.y && this.z === rhs.z && this.w === rhs.w;
-    },
+    }
 
     /**
      * @function
@@ -154,9 +211,9 @@ Object.assign(Vec4.prototype, {
      * // Should output 5
      * console.log("The length of the vector is: " + len);
      */
-    length: function () {
+    length(): number {
         return Math.sqrt(this.x * this.x + this.y * this.y + this.z * this.z + this.w * this.w);
-    },
+    }
 
     /**
      * @function
@@ -169,9 +226,9 @@ Object.assign(Vec4.prototype, {
      * // Should output 25
      * console.log("The length squared of the vector is: " + len);
      */
-    lengthSq: function () {
+    lengthSq(): number {
         return this.x * this.x + this.y * this.y + this.z * this.z + this.w * this.w;
-    },
+    }
 
     /**
      * @function
@@ -192,14 +249,14 @@ Object.assign(Vec4.prototype, {
      * r.lerp(a, b, 0.5); // r is 5, 5, 5, 5
      * r.lerp(a, b, 1);   // r is equal to b
      */
-    lerp: function (lhs, rhs, alpha) {
+    lerp(lhs: Vec4, rhs: Vec4, alpha: number): Vec4 {
         this.x = lhs.x + alpha * (rhs.x - lhs.x);
         this.y = lhs.y + alpha * (rhs.y - lhs.y);
         this.z = lhs.z + alpha * (rhs.z - lhs.z);
         this.w = lhs.w + alpha * (rhs.w - lhs.w);
 
         return this;
-    },
+    }
 
     /**
      * @function
@@ -216,14 +273,14 @@ Object.assign(Vec4.prototype, {
      * // Should output 8, 15, 24, 35
      * console.log("The result of the multiplication is: " + a.toString());
      */
-    mul: function (rhs) {
+    mul(rhs: Vec4): Vec4 {
         this.x *= rhs.x;
         this.y *= rhs.y;
         this.z *= rhs.z;
         this.w *= rhs.w;
 
         return this;
-    },
+    }
 
     /**
      * @function
@@ -242,14 +299,14 @@ Object.assign(Vec4.prototype, {
      * // Should output 8, 15, 24, 35
      * console.log("The result of the multiplication is: " + r.toString());
      */
-    mul2: function (lhs, rhs) {
+    mul2(lhs: Vec4, rhs: Vec4): Vec4 {
         this.x = lhs.x * rhs.x;
         this.y = lhs.y * rhs.y;
         this.z = lhs.z * rhs.z;
         this.w = lhs.w * rhs.w;
 
         return this;
-    },
+    }
 
     /**
      * @function
@@ -265,7 +322,7 @@ Object.assign(Vec4.prototype, {
      * // Should output 1, 0, 0, 0
      * console.log("The result of the vector normalization is: " + v.toString());
      */
-    normalize: function () {
+    normalize(): Vec4 {
         var lengthSq = this.x * this.x + this.y * this.y + this.z * this.z + this.w * this.w;
         if (lengthSq > 0) {
             var invLength = 1 / Math.sqrt(lengthSq);
@@ -276,7 +333,7 @@ Object.assign(Vec4.prototype, {
         }
 
         return this;
-    },
+    }
 
     /**
      * @function
@@ -297,14 +354,14 @@ Object.assign(Vec4.prototype, {
      * // Divide by 2
      * v.scale(0.5);
      */
-    scale: function (scalar) {
+    scale(scalar: number): Vec4 {
         this.x *= scalar;
         this.y *= scalar;
         this.z *= scalar;
         this.w *= scalar;
 
         return this;
-    },
+    }
 
     /**
      * @function
@@ -322,14 +379,14 @@ Object.assign(Vec4.prototype, {
      * // Should output 5, 10, 20, 40
      * console.log("The result of the vector set is: " + v.toString());
      */
-    set: function (x, y, z, w) {
+    set(x: number, y: number, z: number, w: number): Vec4 {
         this.x = x;
         this.y = y;
         this.z = z;
         this.w = w;
 
         return this;
-    },
+    }
 
     /**
      * @function
@@ -346,14 +403,14 @@ Object.assign(Vec4.prototype, {
      * // Should output [-10, -10, -10, -10]
      * console.log("The result of the subtraction is: " + a.toString());
      */
-    sub: function (rhs) {
+    sub(rhs: Vec4): Vec4 {
         this.x -= rhs.x;
         this.y -= rhs.y;
         this.z -= rhs.z;
         this.w -= rhs.w;
 
         return this;
-    },
+    }
 
     /**
      * @function
@@ -372,14 +429,14 @@ Object.assign(Vec4.prototype, {
      * // Should output [-10, -10, -10, -10]
      * console.log("The result of the subtraction is: " + r.toString());
      */
-    sub2: function (lhs, rhs) {
+    sub2(lhs: Vec4, rhs: Vec4): Vec4 {
         this.x = lhs.x - rhs.x;
         this.y = lhs.y - rhs.y;
         this.z = lhs.z - rhs.z;
         this.w = lhs.w - rhs.w;
 
         return this;
-    },
+    }
 
     /**
      * @function
@@ -391,92 +448,28 @@ Object.assign(Vec4.prototype, {
      * // Should output '[20, 10, 5, 0]'
      * console.log(v.toString());
      */
-    toString: function () {
+    toString(): string {
         return '[' + this.x + ', ' + this.y + ', ' + this.z + ', ' + this.w + ']';
     }
-});
 
-/**
- * @field
- * @name pc.Vec4#x
- * @type {number}
- * @description The first component of the vector.
- * @example
- * var vec = new pc.Vec4(10, 20, 30, 40);
- *
- * // Get x
- * var x = vec.x;
- *
- * // Set x
- * vec.x = 0;
- */
-/**
- * @field
- * @name pc.Vec4#y
- * @type {number}
- * @description The second component of the vector.
- * @example
- * var vec = new pc.Vec4(10, 20, 30, 40);
- *
- * // Get y
- * var y = vec.y;
- *
- * // Set y
- * vec.y = 0;
- */
-/**
- * @field
- * @name pc.Vec4#z
- * @type {number}
- * @description The third component of the vector.
- * @example
- * var vec = new pc.Vec4(10, 20, 30, 40);
- *
- * // Get z
- * var z = vec.z;
- *
- * // Set z
- * vec.z = 0;
- */
-/**
- * @field
- * @name pc.Vec4#w
- * @type {number}
- * @description The fourth component of the vector.
- * @example
- * var vec = new pc.Vec4(10, 20, 30, 40);
- *
- * // Get w
- * var w = vec.w;
- *
- * // Set w
- * vec.w = 0;
- */
-
-/**
- * @field
- * @static
- * @readonly
- * @name pc.Vec4.ONE
- * @type {pc.Vec4}
- * @description A constant vector set to [1, 1, 1, 1].
- */
-
-/**
- * @field
- * @static
- * @readonly
- * @name pc.Vec4.ZERO
- * @type {pc.Vec4}
- * @description A constant vector set to [0, 0, 0, 0].
- */
-
-Object.defineProperties(Vec4, {
-    ZERO: { value: new Vec4(0, 0, 0, 0) },
-    ONE: { value: new Vec4(1, 1, 1, 1) }
-});
-
-Object.freeze(Vec4.ZERO);
-Object.freeze(Vec4.ONE);
+    /**
+     * @field
+     * @static
+     * @readonly
+     * @name pc.Vec4.ZERO
+     * @type {pc.Vec4}
+     * @description A constant vector set to [0, 0, 0, 0].
+     */
+    static readonly ZERO: Vec4 = Object.freeze(new Vec4(0, 0, 0, 0));
+    /**
+     * @field
+     * @static
+     * @readonly
+     * @name pc.Vec4.ONE
+     * @type {pc.Vec4}
+     * @description A constant vector set to [1, 1, 1, 1].
+     */
+    static readonly ONE: Vec4 = Object.freeze(new Vec4(1, 1, 1, 1));
+}
 
 export { Vec4 };
